[권지민] 코드래빗 기반 수정
